Separate NewGameWindow into its own file
