+++ src/package.bluej
@@ -1,19 +1,36 @@
 #BlueJ package file
+dependency1.from=VirtualGameShelf
+dependency1.to=NewGameWindow
+dependency1.type=UsesDependency
 objectbench.height=76
 objectbench.width=658
 package.editor.height=400
 package.editor.width=560
 package.editor.x=632
 package.editor.y=225
-package.numDependencies=0
-package.numTargets=1
+package.numDependencies=1
+package.numTargets=2
 package.showExtends=true
 package.showUses=true
+project.charset=windows-1252
 target1.height=50
-target1.name=VirtualGameShelf
+target1.name=NewGameWindow
 target1.showInterface=false
 target1.type=ClassTarget
 target1.typeParameters=
-target1.width=120
-target1.x=160
-target1.y=10
+target1.width=130
+target1.x=200
+target1.y=70
+target2.editor.height=700
+target2.editor.width=900
+target2.editor.x=79
+target2.editor.y=385
+target2.height=50
+target2.name=VirtualGameShelf
+target2.naviview.expanded=true
+target2.showInterface=false
+target2.type=ClassTarget
+target2.typeParameters=
+target2.width=120
+target2.x=70
+target2.y=10
